task_one/python/requirements.txt
kafka-python==2.0.2
avro-python3==1.10.2
requests==2.31.0
six==1.16.0 